# api-docs.yaml
openapi: 3.0.1
info:
  title: Instant Messaging API
  description: |
    API REST pour une messagerie instantanée qui comprend la gestion  des utilisateur·rice·s, des conversations et des messages. Cette API est utilisée dans  le cadre des modules 194, 195, 322, et 426 à l'EPAI Fribourg.
  contact:
    name: Jérôme Frossard
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
  description: Generated server url
security:
- BearerAuth: []
...
paths:
...
  /conversation/{conversationId}/messages:
    get:
      tags:
      - message
      summary: Récupère les messages d'une conversation
      description: Récupère les messages d'une conversation donnée. L'utilisateur·rice
        connecté·e doit être membre de la conversation.
      operationId: get-messages
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "415":
          description: Unsupported Media Type
        "200":
          description: Liste de messages récupérée avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ConversationsReponseDto"
              example:
                _embedded:
                  messages:
                  - id: 100
                    senderId: 11
                    body: Hi Idaho!
                    participantStatus:
                    - userId: 11
                      readAt: null
                      deleted: false
                    - userId: 12
                      readAt: null
                      deleted: false
                    sentAt: 2025-05-07T13:48:31
                    _links:
                      self:
                        href: /api/v1/conversation/100/messages/100
                  - id: 101
                    senderId: 12
                    body: "Hey Sheana, good to hear from you."
                    participantStatus:
                    - userId: 11
                      readAt: null
                      deleted: false
                    - userId: 12
                      readAt: null
                      deleted: false
                    sentAt: 2025-05-07T13:48:31
                    _links:
                      self:
                        href: /api/v1/conversation/100/messages/101
                _links:
                  self:
                    href: /api/v1/conversation/100/messages?page=0&size=20
                page:
                  size: 20
                  totalElements: 2
                  totalPages: 1
                  number: 0
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
    post:
      tags:
      - message
      summary: Ajoute un message à la conversation
      description: Ajoute un message à la conversation à partir de son identifiant.
      operationId: create-message
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreateDto"
            example:
              message: Salut! Quoi de neuf?
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Rôle ou statut de lea participant·modifié avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ParticipantResponseDto"
              example:
                userId: 15
                username: fallom@example.com
                role: MEMBER
                status: INACTIVE
                _links:
                  self:
                    href: http://localhost:8080/api/v1/conversation/110/participants/15
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/MessageResponseDto"
...
  /conversation/{conversationId}/messages/{messageId}:
    patch:
      tags:
      - message
      summary: Modifie le statut du message pour un participant
      description: |
        Modifie le statut d’un message pour l’utilisateur·rice connecté·e.

        * Pour marquer le message comme lu, le champ `read` est mis à true. Si le message a déjà été lu, l'opération n'a pas d'effet.

        * Pour marquer le message comme supprimé, le champ `deleted` est mis à `true`. Cette opération est irréversible et entraîne la suppression logique du message pour cette personne.
      operationId: update-messages-status
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      - name: messageId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageUpdateDto"
            examples:
              Marquer le message comme lu.:
                description: Marquer le message comme lu.
                value:
                  read: true
              Marquer le message comme supprimer.:
                description: Marquer le message comme supprimer.
                value:
                  delete: true
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Statut·modifié avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ParticipantResponseDto"
              example:
                id: 100
                senderId: 11
                body: Hi Idaho!
                participantStatus:
                - userId: 11
                  readAt: 2025-05-07T20:55:17
                  deleted: false
                - userId: 12
                  readAt: null
                  deleted: false
                sentAt: 2025-05-07T20:08:12
                _links:
                  self:
                    href: /api/v1/conversation/100/messages/100
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/Link"
...
    ConversationResponseDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            participants:
              type: array
              items:
                $ref: "#/components/schemas/ParticipantResponseDto"
        _links:
          $ref: "#/components/schemas/Links"
        id:
          type: integer
          format: int64
        group:
          type: boolean
    ParticipantResponseDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
        status:
          type: string
        _links:
          $ref: "#/components/schemas/Links"
    MessageCreateDto:
      type: object
      properties:
        message:
          type: string
    MessageResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senderId:
          type: integer
          format: int64
        body:
          type: string
        sentAt:
          type: string
          format: date-time
        participantStatus:
          type: array
          items:
            $ref: "#/components/schemas/MessageStatusDto"
        _links:
          $ref: "#/components/schemas/Links"
    MessageStatusDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        readAt:
          type: string
          format: date-time
        deleted:
          type: boolean
...
    MessageUpdateDto:
      type: object
      properties:
        read:
          type: boolean
        delete:
          type: boolean
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    ConversationsReponseDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            conversations:
              type: array
              items:
                $ref: "#/components/schemas/ConversationResponseDto"
        _links:
          $ref: "#/components/schemas/Links"
        page:
          $ref: "#/components/schemas/PageMetadata"
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer